Project 1 -  Credibility Score for articles/Deliverable 3/requirements.txt
# requirements.txt

streamlit
openai
trafilatura
beautifulsoup4
textblob
requests
python-dotenv
pytest
lxml
lxml_html_clean